+++ rocketmq-store/src/lib.rs
@@ -30,7 +30,7 @@ pub mod log_file;
 pub(crate) mod message_encoder;
 pub mod message_store;
 pub mod pop;
-mod queue;
+pub mod queue;
 pub(crate) mod services;
 pub mod stats;
 pub mod store;
